==> src_rmd/1_Data_processing.Rmd <==
---
title: "Data processing"
author: "Koya Hashimoto"
date: "2021/01/20"
output:
  md_document:
    variant: markdown_github
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Data processing

### ver 4.1 Last modified: 2024/1/23 minor revision (corrected ZooplaCr scaling)
### ver 4. modified: 2023/8/24 major revision (tentative)

## Load packages

```{r set-up}
sessionInfo() #save session information (R version 3.6.3 (2020))

library(dplyr); packageVersion("dplyr") #v.1.0.2
library(tidyr); packageVersion("tidyr") #v.1.1.2
library(RColorBrewer); packageVersion("RcolorBrewer") #v.1.1.2
library(ggplot2); packageVersion("ggplot2") #v.3.3.2
library(ggsci); packageVersion("ggsci") #v.2.9
library(patchwork); packageVersion("patchwork") #v.1.1.1
```

## Data compile

- Cont: Control
- Fipro: Fipronil (insecticide) only
- Pent: Pentoxazone (herbicide) only
- Joint: Fipronil+Pentoxazone
- Exclude -1 and 1 weeks 

```{r  data, fig.width=10, fig.height=4}
PlanktonData <- read.csv("./raw_data/PlaData171819_corrected.csv", header=TRUE)
MacrophyteData <- read.csv("./raw_data/MacrophyteData171819.csv", header=TRUE)
AnimalData <- read.csv("./raw_data/animaldata171819.completegathered.fixed.csv", header=TRUE)

# Data compilation

data_definefctr <- function(data) {
  data$Insecticide <- as.factor(data$Insecticide)
  data$Herbicide <- as.factor(data$Herbicide)
  data$Treatment <- factor(data$Treatment, levels=c("Cont", "Fipro", "Pent", "Joint")) #Define category order
  data$Tank <- factor(data$Tank, levels=c("Cont1", "Cont2", "Fipro1", "Fipro2", "Pent1", "Pent2", "Joint1", "Joint2"))
  data$Year_Tank <- with(data, paste(Year, Tank, sep="_")) #Define unit of analyses
  data$Year_Tank <- 
    factor(data$Year_Tank, 
           levels=c("2017_Cont1", "2017_Cont2", "2017_Fipro1", "2017_Fipro2", 
                    "2017_Pent1", "2017_Pent2", "2017_Joint1", "2017_Joint2", 
                    "2018_Cont1", "2018_Cont2", "2018_Fipro1", "2018_Fipro2", 
                    "2018_Pent1", "2018_Pent2", "2018_Joint1", "2018_Joint2", 
                    "2019_Cont1", "2019_Cont2", "2019_Fipro1", "2019_Fipro2", 
                    "2019_Pent1", "2019_Pent2", "2019_Joint1", "2019_Joint2")) #Define category order
  return(data)
}

PlanktonData <- data_definefctr(PlanktonData)
MacrophyteData <- data_definefctr(MacrophyteData)
MacrophyteData$SubPlot <- as.factor(MacrophyteData$SubPlot)
AnimalData <- data_definefctr(AnimalData)

# Split by functional groups

Phytopla <- 
  PlanktonData %>%
  filter(Group=="Diatom" | Group=="Flagellate" | Group=="GreenAlgae") %>% # Omit unnecessary taxa
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Roti <- 
  PlanktonData %>%
  filter((Group=="Rotifer") & Species_each_file!="Nauplius_(E._japonicus)") %>%  #omit a different sampling protocol
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
#Miki Hirao counted all individuals in 1L sample, whereas Ji Cai took 1/10 subsample from 500mL sample and counted individuals in the 1/10 subsample
ZooplaCr <- 
  PlanktonData %>%
  filter((Group=="Cladocera" | Group=="Copepoda" | Group=="Ostracoda") & Species_each_file!="Nauplius_(E._japonicus)") %>%  #omit a different sampling protocol
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Mp <- 
  MacrophyteData %>%
  group_by(Year, Tank, Week, SubPlot, Year_Tank) %>%
  summarise(Cover=sum(Cover)) %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=mean(Cover)) %>%
  ungroup(.) %>% filter(Week>1)
Det <- 
  AnimalData %>%
  filter(Function=="Detritivore") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Herb <- 
  AnimalData %>%
  filter(Function=="Herbivore") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Pred <- 
  AnimalData %>%
  filter(Function=="Predator") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)

Pred.P <- 
  AnimalData %>%
  filter(Function=="Predator" & Habitat=="Phytophilous") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Pred.B <-   AnimalData %>%
  filter(Function=="Predator" & Habitat=="Benthic") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Pred.S <-   AnimalData %>%
  filter(Function=="Predator" & Habitat=="Neustonic") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)
Pred.C <-   AnimalData %>%
  filter(Function=="Predator" & Habitat=="Nektonic") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)

Mollusca <- 
  AnimalData %>%
  filter(Species=="Bivalvia_sp" | Species=="Physa_acuta") %>%
  group_by(Year, Tank, Week, Year_Tank) %>%
  summarise(Abundance=sum(Abundance)) %>%
  ungroup(.) %>% filter(Week>1)

Commulist <- list(Phytopla, Roti, ZooplaCr, Mp, Det, Herb, Pred.P, Pred.B, Pred.S, Pred.C, Mollusca)
names(Commulist) <- c("Phytopla", "Roti", "ZooplaCr", "Mp", "Det", "Herb", "Pred.P", "Pred.B", "Pred.S", "Pred.C", "Mollusca")

# Display raw time series

theme_set(theme_bw())
lapply(1:3, function(i) {
  Commulist[[i]] %>%
    separate(Tank, into=c("Treatment", "Num"), sep=-1) %>% 
    mutate(Treatment=factor(Treatment, levels=c("Cont", "Fipro", "Pent", "Joint"))) %>%
    ggplot(aes(x=Week, y=log10(Abundance+1), color=Treatment, linetype=Num)) + geom_line(size=1) + facet_wrap(~Year) + ggtitle(names(Commulist)[i])
}
)

lapply(4:11, function(i) {
  Commulist[[i]] %>%
    separate(Tank, into=c("Treatment", "Num"), sep=-1) %>% 
    mutate(Treatment=factor(Treatment, levels=c("Cont", "Fipro", "Pent", "Joint"))) %>%
    ggplot(aes(x=Week, y=Abundance, color=Treatment, linetype=Num)) + geom_line(size=1) + facet_wrap(~Year) + ggtitle(names(Commulist)[i])
}
)

# Log-transformation and standardization

Tslist <- lapply(1:11, function(i) {
  Commulist[[i]] %>%
    select(Week, Year_Tank, Abundance) %>%
    spread(key=Year_Tank, value=Abundance) %>% # Remove unnecessary columns
    rbind(., rep(NaN, ncol(.))) %>% # Add NaN to separate different tanks, but not necessary if you correctly specify lib parameters
    gather(Year_Tank, Abundance, -Week) %>% # Gather data for EDM
    mutate(Year_Tank=factor(Year_Tank, levels=c("2017_Cont1", "2017_Cont2", "2017_Fipro1", "2017_Fipro2", 
                                                "2017_Pent1", "2017_Pent2", "2017_Joint1", "2017_Joint2", 
                                                "2018_Cont1", "2018_Cont2", "2018_Fipro1", "2018_Fipro2", 
                                                "2018_Pent1", "2018_Pent2", "2018_Joint1", "2018_Joint2", 
                                                "2019_Cont1", "2019_Cont2", "2019_Fipro1", "2019_Fipro2", 
                                                "2019_Pent1", "2019_Pent2", "2019_Joint1", "2019_Joint2")))
  } 
)

names(Tslist) <- 
  c("Phytopla1", "Roti1", "Zoopla1", "Mp1", "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Pred.C1", "Moll1")

lapply(c("Phytopla1", "Roti1", "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Pred.C1", "Moll1"), 
       function(i) { 
         Tslist[[i]] <<- mutate(Tslist[[i]], scAbundance=as.numeric(scale(log(Abundance+1))))
       }
)

Tslist[["Mp1"]] <- mutate(Tslist[["Mp1"]], scAbundance=as.numeric(scale(Abundance))) # For macrophytes, log-transformation was not done

# Scaling density data of Crustacean zooplankton because different sampling protocols were used in different years

par(mfrow=c(1, 2))
boxplot(I(Abundance+1) ~ Year, ZooplaCr)
Tslist[["Zoopla1"]] <- 
  Tslist[["Zoopla1"]] %>%
  separate(Year_Tank, into=c("Year", "Tank")) %>%
  group_by(Year) %>%
  mutate(scAbundance=as.numeric(scale(log(Abundance*ifelse(Year==2017, 1, 20)+1)))) %>% # To account for different sampling protocols between different years
  unite(Year, Tank, col="Year_Tank")
Tslist[["Zoopla1"]] %>%
  separate(Year_Tank, into=c("Year", "Tank")) %>%
  boxplot(scAbundance ~ Year, .)

all_Ts <- 
  lapply(1:11, function(i) {
    mutate(Tslist[[i]], Recipient=names(Tslist)[i])
  }) %>% do.call(rbind, .)
write.csv(all_Ts, "./processed_data/all_Ts240123.csv", row.names=FALSE) # Save processed time series data
```

## Data visualization

### Processing for plotting

```{r data2}
#Abundance.gathered <- read.csv("./processed_data/all_Ts240123.csv", header=TRUE)

Abundance.gathered <- all_Ts

Abundance.gathered_ <- 
  Abundance.gathered %>%
  separate(Year_Tank, into=c("Year", "Tank"), sep="_") %>%
  mutate(Treatment=substr(Tank, start=1, stop=nchar(Tank)-1)) %>%
  mutate(Treatment=factor(Treatment, levels=c("Cont", "Fipro", "Pent", "Joint"))) %>%
  unite(col="Treatment_Tank_Recipient", Treatment, Tank, Recipient, sep="_") %>%
  select(-Abundance) %>%
  spread(key="Treatment_Tank_Recipient", value=scAbundance) %>%
  arrange(Year, Week) %>%
  rbind(., NaN) %>%
  gather(key="Treatment_Tank_Recipient", value="scAbundance", -(1:2)) %>%
  separate(Treatment_Tank_Recipient, into=c("Treatment", "Tank", "Recipient"), sep="_") %>%
  unite(col="Treatment_Recipient", Treatment, Recipient, sep="_") %>% 
  mutate(Tank=substr(Tank, nchar(Tank), nchar(Tank))) %>%
  spread(key="Treatment_Recipient", value="scAbundance") %>% 
  arrange(Tank, Year, Week) %>%
  rbind(., NaN) %>%
  gather(key="Treatment_Recipient", value="scAbundance", -(1:3)) %>%
  separate(Treatment_Recipient, into=c("Treatment", "Recipient"), sep="_") %>%
  unite(col="Tank", Treatment, Tank, remove=FALSE, sep="") %>%
  mutate(Treatment=factor(Treatment, levels=c("Cont", "Fipro", "Pent", "Joint")), 
         Tank=factor(Tank, levels=c("Cont1", "Cont2", "ContNaN", "Fipro1", "Fipro2", "FiproNaN",
                                    "Pent1", "Pent2", "PentNaN", "Joint1", "Joint2", "JointNaN")), 
         Year=as.numeric(Year)) %>%
  mutate(Recipient=factor(Recipient, levels=c("Phytopla1", "Roti1", "Zoopla1", "Mp1", "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Pred.C1", "Moll1"))) %>%
  arrange(Recipient, Tank, Year, Week)


map <- c("Phytopla1"="Phytoplankton", 
          "Roti1"="Rotifers", 
          "Zoopla1"="Crustacean zooplankton", 
          "Mp1"="Macrophytes", 
          "Det1"="Detritivores", 
          "Herb1"="Herbivores", 
          "Pred.P1"="Phytophilous predators", 
          "Pred.B1"="Benthic predators",
          "Pred.S1"="Neustonic predators", 
          "Pred.C1"="Nektonic predators",
          "Moll1"="Molluscs")

Abundance.gathered_mod <- 
  Abundance.gathered_  %>%
  mutate(Recipient=factor(map[Recipient], 
                      levels=c("Phytoplankton", 
                               "Rotifers", 
                               "Crustacean zooplankton", 
                               "Macrophytes", 
                               "Detritivores", 
                               "Herbivores", 
                               "Phytophilous predators", 
                               "Benthic predators", 
                               "Neustonic predators", 
                               "Molluscs")), 
         Time=rep(1:69, times=nrow(.)/69), 
         Time2=rep(1:276, times=nrow(.)/276)) %>% 
  filter(Recipient!="Pred.C1")
  
g_Aball <- ggplot(Abundance.gathered_mod, aes(x=Time2, y=scAbundance)) +
  geom_line(aes(color=Recipient)) +
  facet_wrap(~Recipient, ncol=2, scales="free", labeller=label_value) + 
  scale_y_continuous(limits=c(-3.1, 3.1), name="Standardized density") +
  scale_x_continuous(name=NULL, breaks=seq(276/48, by=276/24, length=24)-1, 
                     labels=rep(c("'17", "'18", "'19"), times=8), limits=c(1, 276), 
                     sec.axis=sec_axis(~.,
                                       breaks=seq(276/8, by=276/4, length=4)-1, 
                                       labels=c("C", "I", "H", "I+H"))) + 
  theme_test() +
  theme(legend.position="none", 
        axis.ticks.x=element_blank(), strip.background=element_blank(), 
        strip.text.x=element_text(hjust=0), 
        strip.placement="outside") + 
  geom_hline(yintercept=0, linetype=2) + 
  geom_vline(xintercept=seq(276/8, by=(276/8)*2, length=4)-1, color="grey") + 
  geom_vline(xintercept=seq(276/4, by=(276/8)*2, length=3)-1, color="black") +
  scale_color_manual(values=(pal_d3(palette="category10")(10))[c(5:9, 1:4, 10)])
```

### Fig. S2a

```{r fig, fig.width=12, fig.height=7.5}
#windows(12, 7.5, rescale="fixed")
g_Aball
#ggsave("fig.s3.pdf", fig.s3, device=cairo_pdf, width=12, height=15, unit="in")
```
