==> src_rmd/3_Calculation_of_population_sensitivity.Rmd <==
---
title: "Calculation of population sensitivity"
author: "Koya Hashimoto"
date: "2024/6/14"
output:
  md_document:
    variant: markdown_github
---

# Calculation of population sensitiity

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

### Load required packages
  
```{r packages}
sessionInfo() #save session information (R version 3.6.3 (2020))

library(ggplot2); packageVersion("ggplot2") #3.3.2
library(dplyr); packageVersion("dplyr") #1.0.2
library(tidyr); packageVersion("tidyr") #1.1.2
library(patchwork); packageVersion("patchwork") #1.1.1
library(RColorBrewer); packageVersion("RColorBrewer") #1.1.2
library(ggsci); packageVersion("ggsci") #2.9
```

### Data compilation

```{r data compile}

# Load standardized time series

all_Ts <- read.csv("./processed_data/all_Ts240123.csv") 

all_Ts <- #add required columns
  all_Ts %>%
  separate(Year_Tank, into=c("Year", "Tank"), sep="_") %>% #year and tank
  mutate(Treatment=substr(Tank, 1, nchar(Tank)-1)) %>% #treatment name
  mutate(Treatment=factor(Treatment, levels=c("Cont", "Fipro", "Pent", "Joint")), #order of treatments
         Tank=factor(Tank, levels=unique(Tank)), #order of tanks
         Recipient=factor(Recipient, levels=c("Phytopla1", "Roti1", "Zoopla1", "Mp1", 
                                              "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Moll1"))) #order of recipient populations
```

## Calculation of population sensitivity (with Fig. 2b)

```{r calc sensitiivty}

# Averaging density over replicates

all_Ts_smrz <- 
  all_Ts %>%
  group_by(Treatment, Recipient, Year) %>%
  summarise(Abundance_m=mean(Abundance, na.rm=TRUE)) #mean raw density


## Show data sheet

head(all_Ts_smrz)

# Calculating log response ratio

Abundance_TrEff_nRR_0.1 <-  # calculate log response ratio following the method from Mratinson and Raupp 2013
  all_Ts_smrz %>%
  ungroup(.) %>%
  spread(value=Abundance_m, key=Treatment) %>%
  mutate(Fipro_Cont=log((Fipro+0.1)/(Cont+0.1)), 
         Pent_Cont=log((Pent+0.1)/(Cont+0.1)),
         Joint_Cont=log((Joint+0.1)/(Cont+0.1))) %>%
  select(-Cont, -Fipro, -Pent, -Joint) %>%
  gather(key=Treatment_Cont, value=Abundance_response_nRR_0.1, -(1:2)) %>%
  mutate(Treatment_Cont=factor(Treatment_Cont, levels=c("Fipro_Cont", "Pent_Cont", "Joint_Cont")))

Abundance_TrEff <- Abundance_TrEff_nRR_0.1
Abundance_TrEff <- filter(Abundance_TrEff, Recipient!="Pred.C1") # omit nektonic predators
Abundance_TrEff$abs_Abundance_response_nRR_0.1 <- abs(Abundance_TrEff$Abundance_response_nRR_0.1) #Absolute values

Abundance_TrEff$Recipient <- factor(Abundance_TrEff$Recipient, 
                                    levels=c("Phytopla1", "Roti1", "Zoopla1", "Mp1", 
                                             "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Moll1"))

## Show data sheet

head(Abundance_TrEff)

# Visualize sensitivity

smrz_Abundance_TrEff <- 
  Abundance_TrEff %>%
  group_by(Recipient, Treatment_Cont) %>%
  summarise(abs_Abundance_response_nRR_0.1=(exp(mean(log(abs_Abundance_response_nRR_0.1+1)))-1)) %>% #calculating geometric mean
  ungroup(.)

## Split by treatments

theme_custom <- function() {
  theme_bw() + 
    theme(panel.border=element_rect(fill=NA, 
                                    colour="black"), 
          panel.spacing=unit(3, "pt"),
          strip.text=element_text(colour="black", lineheight=0.7))
}

palSet1 <- brewer.pal(9, "Set1")
reaction.norm <-
  smrz_Abundance_TrEff %>%
  #  mutate(Treatment_Cont=factor(Treatment_Cont, levels=c("Fipro_Cont", "Joint_Cont", "Pent_Cont"))) %>%
  ggplot(aes(x=Recipient, y=log(abs_Abundance_response_nRR_0.1+1))) +
  geom_point(aes(group=Treatment_Cont, color=Treatment_Cont, shape=Treatment_Cont), size=4) +
  geom_line(aes(group=Treatment_Cont, color=Treatment_Cont)) +
  geom_point(Abundance_TrEff, 
             mapping=aes(x=Recipient, y=log(abs_Abundance_response_nRR_0.1+1), 
                         group=Treatment_Cont, 
                         color=Treatment_Cont, 
                         shape=Treatment_Cont), 
             alpha=0.4, size=2.5) +
  scale_color_manual(values=palSet1[c(1, 2, 3)], labels=c("I vs. C", "H vs. C", "I+H vs. C")) + 
  scale_shape_manual(values=c(16, 17, 15), labels=c("I vs. C", "H vs. C", "I+H vs. C")) + 
  scale_x_discrete(name="Community member", labels=c("Phyt", "Roti", "C.zoop", "Macr", "Detr", "Herb", "P.pred", "B.pred", "N.pred", "Moll")) + 
  guides(color=guide_legend(title=NULL), shape=guide_legend(title=NULL)) +
  ylab("ln(Population sensitivity + 1)") + 
  theme_custom()

reaction.norm #Fig. 2b

write.csv(Abundance_TrEff, "./processed_data/Abundance_TrEff.csv", row.names=FALSE) # export sensitivity data
```

## Generate Fig. 2

```{r fig 2a}
# Visualize data

labeli <- as_labeller(c(`Phytopla1`="Phytoplankton", #map facet labels
                        `Roti1`="Rotifers",
                        `Zoopla1`="Crustacean\nzooplankton",
                        `Mp1`="Macrophytes",
                        `Det1`="Detritivores",
                        `Herb1`="Herbivores",
                        `Pred.P1`="Phytophilous\npredators",
                        `Pred.B1`="Benthic\npredators",
                        `Pred.S1`="Neustonic\npredators",
                        `Moll1`="Molluscs"))
lab <- data.frame(lab="*", 
                  Recipient=factor(c("Phytopla1", "Mp1", "Mp1", 
                                 "Herb1", "Pred.P1", "Pred.P1", 
                                 "Pred.B1", "Pred.B1", "Moll1"), 
                               levels=c("Phytopla1", "Roti1", "Zoopla1", "Mp1", 
                                        "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Moll1")))

pval <- data.frame(lab=c("0.62", "0.92", "0.02", 
                         "0.62", "0.97", "0.78", 
                         "0.65", "0.33", "0.95", 
                         "0.98", "0.02", "0.0004", 
                         "0.46", "0.82", "0.84", 
                         "0.86", "0.14", "0.0007", 
                         "0.0001", "0.12", "0.0001", 
                         "<0.0001", "0.99", "<0.0001", 
                         "0.44", "0.91", "0.99", 
                         "0.49", "0.96", "0.003"), 
                  Recipient=factor(rep(c("Phytopla1", "Roti1", "Zoopla1", "Mp1", 
                                     "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Moll1"), each=3), 
                                   levels=c("Phytopla1", "Roti1", "Zoopla1", "Mp1", 
                                            "Det1", "Herb1", "Pred.P1", "Pred.B1", "Pred.S1", "Moll1")))

palSet1 <- brewer.pal(9, "Set1")


set.seed(1)
density.boxplot_gg <- 
  ggplot(na.omit(filter(all_Ts, Recipient!="Pred.C1")), mapping=aes(x=Treatment, y=scAbundance)) +
  geom_boxplot(mapping=aes(fill=Treatment), outlier.shape=NA, size=0.3, colour="black") + 
  scale_fill_manual(values=palSet1[c(9, 1, 2, 3)]) + 
  geom_jitter(width=0.2, size=0.1) + 
  facet_wrap(~Recipient, nrow=2, labeller=labeli) + 
  scale_x_discrete(labels=c("C", "I", "H", "I+H")) +
  theme_custom() + #theme
  theme(strip.background=element_blank(), legend.position="none") +
  geom_text(data=lab, aes(label=lab), x=c(4, 3, 4, 4, 2, 4, 2, 4, 4), y=2.3, size=10) +
  geom_text(data=pval, aes(label=lab), x=rep(2:4, times=10), y=3.5, size=2.2) +
  labs(y="Standardized density") #軸ラベル
```

```{r paper fig, fig.width=8, fig.height=7}
#windows(8, 7, rescale="fixed")
density.boxplot_gg / reaction.norm + plot_layout(height=c(5, 3)) + plot_annotation(tag_levels="a") & 
  theme(plot.tag=element_text(face="bold"), text=element_text(size=14))
ggsave("./figs/fig2.ggplot2.pdf", width=8, height=7, device=cairo_pdf, unit="in")
```
